refactor(user): migrate registration service from auth to user module
